CreateVectorStore-Service/requirements.txt
fastapi
uvicorn
PyMuPDF
langchain
sentence-transformers
langchain-google-genai
langchain-openai
langchain-experimental
langchain-community
pymilvus
python-dotenv
redis
icalendar
pyjwt
twilio
motor
pymongo
boto3
email-validator
pydantic
python-multipart
validators
bs4
unstructured
openpyxl
docx2txt
firecrawl-py
firecrawl
selenium
webdriver-manager
html2text
langchain-pinecone
pinecone-client